Update rename packages, use public PyPI artifacts
