+++ setup.cfg
@@ -29,7 +29,7 @@ include_trailing_comma = true
 combine_as_imports = True
 force_grid_wrap = 0
 known_first_party = rpdk
-known_third_party = boto3,jinja2,aws_cloudformation_rpdk_python_lib,pytest
+known_third_party = boto3,jinja2,cloudformation_cli_python_lib,pytest
 
 [tool:pytest]
 # can't do anything about 3rd party modules, so don't spam us
