+++ src/cloudformation_cli_python_lib/exceptions.py
