+++ src/cloudformation_cli_python_lib/resource.py
