+++ src/cloudformation_cli_python_lib/utils.py
