+++ src/setup.py
@@ -2,15 +2,18 @@
 from setuptools import setup
 
 setup(
-    name="aws-cloudformation-rpdk-python-lib",
+    name="cloudformation-cli-python-lib",
     version="0.0.1",
     description=__doc__,
     author="Amazon Web Services",
-    packages=["aws_cloudformation_rpdk_python_lib"],
+    author_email="aws-cloudformation-developers@amazon.com",
+    url="https://github.com/aws-cloudformation/aws-cloudformation-rpdk-python-plugin/",
+    packages=["cloudformation_cli_python_lib"],
     # package_data -> use MANIFEST.in instead
     include_package_data=True,
     zip_safe=True,
-    install_requires=["boto3>=1.9.108,<1.10", 'dataclasses;python_version<"3.7"'],
+    python_requires=">=3.6",
+    install_requires=["boto3>=1.10.20", 'dataclasses;python_version<"3.7"'],
     license="Apache License 2.0",
     classifiers=[
         "Development Status :: 2 - Pre-Alpha",
