+++ tests/lib/boto3_proxy_test.py
@@ -1,8 +1,5 @@
-from aws_cloudformation_rpdk_python_lib.boto3_proxy import (
-    SessionProxy,
-    _get_boto_session,
-)
-from aws_cloudformation_rpdk_python_lib.utils import Credentials
+from cloudformation_cli_python_lib.boto3_proxy import SessionProxy, _get_boto_session
+from cloudformation_cli_python_lib.utils import Credentials
 
 
 def test_get_boto_session_returns_proxy():
